Change version field size to int16

+++ btcjson/chainsvrresults.go
@@ -631,7 +631,7 @@ type TxRawResult struct {
 	Size          int32  `json:"size,omitempty"`
 	Vsize         int32  `json:"vsize,omitempty"`
 	Weight        int32  `json:"weight,omitempty"`
-	Version       int32  `json:"version"`
+	Version       int16  `json:"version"`
 	LockTime      uint32 `json:"locktime"`
 	Vin           []Vin  `json:"vin"`
 	Vout          []Vout `json:"vout"`
@@ -650,7 +650,7 @@ type SearchRawTransactionsResult struct {
 	Size          string       `json:"size"`
 	Vsize         string       `json:"vsize"`
 	Weight        string       `json:"weight"`
-	Version       int32        `json:"version"`
+	Version       int16        `json:"version"`
 	LockTime      uint32       `json:"locktime"`
 	Vin           []VinPrevOut `json:"vin"`
 	Vout          []Vout       `json:"vout"`
@@ -663,7 +663,7 @@ type SearchRawTransactionsResult struct {
 // TxRawDecodeResult models the data from the decoderawtransaction command.
 type TxRawDecodeResult struct {
 	Txid     string `json:"txid"`
-	Version  int32  `json:"version"`
+	Version  int16  `json:"version"`
 	Locktime uint32 `json:"locktime"`
 	Vin      []Vin  `json:"vin"`
 	Vout     []Vout `json:"vout"`
